requirements.txt
langchain==0.0.288
openai==0.28.0
tiktoken==0.5.1
lxml==4.9.3

fastapi==0.103.1
python-dotenv==1.0.0

beautifulsoup4==4.12.2

requests==2.31.0
xmltodict==0.13.0
Unidecode~=1.3.6
pymongo==4.5.0
pydantic==2.3.0
nest_asyncio==1.5.7
uvicorn==0.23.2